Refactor UserAvatar and utils; replace avatar color generation with uniqueColorGenerator utility and enhance color selection logic
